IMPALA-13235: Update Opentelemetry C++ SDK Build Options
The BUILD_SHARED_LIBS option was set to OFF as using the shared
library was forcing that library to be added to LD_LIBRARY_PATH.

The WITH_THREAD_INSTRUMENTATION_PREVIEW option was set to ON as this
experimental option provides hooks for logging detailed information
about the internal threads of the OTLP HTTP exporter. Using this
feature requires adding specific code to activate it.

Change-Id: I46c53184204a0feb87b391ff7fe949bdb2812c58